Bgbot/requirements.txt
python-dotenv
sentence_transformers~=2.2.2
weaviate-client
uvicorn
fastapi
llama_index==0.9.14.post1